% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneTonicWorkshop-package.R
\docType{package}
\name{GeneTonicWorkshop-package}
\alias{GeneTonicWorkshop}
\alias{GeneTonicWorkshop-package}
\title{GeneTonicWorkshop: GeneTonic workshop}
\description{
This workshop covers basic usage of GeneTonic for combining and
    interpreting the results from RNA-seq experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://federicomarini.github.io/GeneTonicWorkshop/}
  \item Report bugs at \url{https://github.com/federicomarini/GeneTonicWorkshop/issues/new/choose}
}

}
\author{
\strong{Maintainer}: Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})

Authors:
\itemize{
  \item Annekathrin Ludt \email{anneludt@uni-mainz.de} (\href{https://orcid.org/0000-0002-2475-4945}{ORCID})
}

}
\keyword{internal}
